Add files via upload
My edits are at the bottom after the code where all the model scores are printed out. 
I added a bar graph of the scores
I tried to do a comparison of overtime and incentive versus actual productivity
I tried Hyper Parameter Tuning of some of the models to see if we can improve the scores.