Add files via upload
I added Regression_Model_Pipeline.py and Regression_Model_Project2.ipynb files. They both are related and need to be together in the same folder to run.
I did the clean, preprocess, and running 9 different regression models and printing out the performance metrics of each. Then there are some graphs showing the metrics. Then a GridSearch function is ran to find the best parameters for each model and then with with the best parameters, the regression models are run again to see if there are changes in the metrics.