Add files via upload
I've updated both the Regression Model Project 2 and Regression Model Pipeline files. I made changes where the training data of just the numerical columns are fitted and transformed while the test data is just transformed. Graphs and charts are updated. The tables are sorted descending based on the MAE which is often preferred when you want a straightforward interpretation of average error. A lower MAE value indicates better model performance, as it means that the model's predictions are closer to the actual values on average. This is generally desirable in predictive modeling. A higher MAE value suggests that the model's predictions are further away from the actual values, indicating poorer performance.